refactor(scripts): share the GitHub client and logger across automation
The stale pull request sweep owned the only hardened GitHub REST client in
the repository: platform fetch, retry classification that tells a rate limit
apart from a denial, and an opt-out for non-idempotent calls. The ownership
gate needs exactly that, and duplicating it would leave two retry policies to
keep in step.

Move the client to scripts/github/client.mjs and the level-aware logger to
scripts/logger.mjs, alongside the other shared script modules. The logger now
takes its prefix from the caller instead of hard-coding one sweep's name.
